Revert "Use shell=True with popen_filtered"
This reverts commit 714b5797262a1866fda1ca24f495798a667fc1e7.

Using 'shell=True' leads to controllers not being killed correctly by
STS. This can lead to errors with MCS etc.

Code that relies on shell expressions should use something along the
lines of
/bin/bash -c 'if [ $all_ok ]; then exec ./controller; fi '

Note the 'exec': It is important that the main controller process
is spawned as the primary executable.